requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pandas>=2.2.0
numpy==2.3.2
python-binance==1.0.19
praw==7.7.1
aiohttp==3.9.1
PyYAML==6.0.1
python-multipart==0.0.6
pytest==8.1.1
notion-client==2.2.1
slack-sdk==3.27.0
python-dotenv==1.0.0
websockets==12.0
httpx==0.27.0
redis==5.0.1
pybreaker==0.4.5
asyncio-mqtt==0.16.1
sentry-sdk==1.40.5
